# src/main/resources/static/docs/project.swagger.yaml
openapi: 3.0.3
info:
  title: Notice Application API
  description: API for the Local Notice Hexagonal WebFlux application
  version: 0.0.1-SNAPSHOT
  contact:
    name: Hexagonal Notice Team
    email: support@hexagonal-notice.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.hexagonal-notice.com
    description: Production server

tags:
  - name: Auth
    description: Authentication operations
  - name: Users
    description: User management operations

paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate user
      description: Authenticate a user and return a JWT token
      operationId: authenticate
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users in the system
      operationId: getAllUsers
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user's information
      operationId: updateUser
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user from the system
      operationId: deleteUser
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
          example: 1
        username:
          type: string
          description: User's unique username
          example: user
        status:
          type: boolean
          description: User account status (active/inactive)
          example: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
          example: "2024-01-15T10:30:00Z"
      required:
        - id
        - username

    AuthenticationRequest:
      type: object
      properties:
        username:
          type: string
          description: User's username
          example: user
        password:
          type: string
          description: User's password
          example: 123456
          format: password
      required:
        - username
        - password

    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      required:
        - token

    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: User's unique username
          example: user
          minLength: 3
          maxLength: 50
        password:
          type: string
          description: User's password
          example: 123456
          minLength: 6
          maxLength: 100
          format: password
      required:
        - username
        - password

    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: User's unique username
          example: user
          minLength: 3
          maxLength: 50
        password:
          type: string
          description: User's password
          example: 123456
          minLength: 6
          maxLength: 100
          format: password
        status:
          type: boolean
          description: User account status
          example: true
      required:
        - username
        - password

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid input provided"
        path:
          type: string
          description: Request path
          example: "/api/users"
      required:
        - timestamp
        - status
        - error
        - message
        - path

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT